test(trino): cover stale metastore ordinals and merge-path pushdown
